Allow the init script to clone from a custom location.
I can now override the source of the `git clone` command. Once cloned,
the URLs in confgit are normalized - pull from Github via https (for
cronjob), push to Github via ssh. This is going to be useful for the
docker build process, where I'll pull from a local directory populated
from the build context.

Also, change the last conditional bit to a proper `if` sequence, so it
doesn't mutilate the exit status.