Switch to path context.
Without explicit `context` specified, `docker/build-push-action@v2`
defaults to a remote context. That fetches a repo in a detached head
state, and the init script had problems cloning from that and
establishing a workable branch that can be used later for subsequent
pulls from remote. While it was possible to change `init.sh` to account
for that, I decided not to rely on the default in case it changes later.
Specifying `context: .` requires a checkout beforehand, but it's
a shallow checkout anyway, so the difference in speed is minimal, and
I don't need to put excessive logic in `init.sh`.