(zsh) Don't run the conf updater if the marker file doesn't exist.
The marker won't be present in environments such as containers. If it is
not there, we won't do anything. The rationale is, containers can be
spun up for a specific short-lived purpose (eg.
https://github.com/yacoob/docker-beets) and we don't really need to run
these with the newest and greatest config files.

A single successful manual run of `conf-dir-update` will create the
marker and enable the check.