(starship) Two lines? Why not three? 😺
In p10k I used to keep the input line all clear, except from the initial
character, and put the timestamp on the right. In starship, the RPROMPT
is vertically aligned with the input line, so it'd create problems with
copying a multi-line command: lots of padding between the command itself
and timestamp on the right.

Other notes:
- `$all` in `format` means "all modules that you haven't explicitly
  mentioned in `format` or `right_format`"
- I've added NF symbols produced by the starship nerd-font-symbols
  preset.