zsh: fix zsh-completions initialization
Antidote was called past compinit, as fzf-tab requires this order. But
that also meant zsh-completions was ineffective, as it needs to be
sourced before compinit is called.

Solution: https://github.com/mattmc3/ez-compinit which "delays" compinit
till end of zshrc, allowing plugins to add things to fpath and use
functions that are defined past compinit.